refactor: 문의 도메인 Inquire로 통일
