+++ Sangwan/src/main/java/com/example/umc10th/domain/inquire/controller/InquireController.java
@@ -1,0 +1,4 @@
+package com.example.umc10th.domain.inquire.controller;
+
+public class InquireController {
+}
