+++ Sangwan/src/main/java/com/example/umc10th/domain/inquire/converter/InquireConverter.java
@@ -1,0 +1,4 @@
+package com.example.umc10th.domain.inquire.converter;
+
+public class InquireConverter {
+}
