+++ Sangwan/src/main/java/com/example/umc10th/domain/inquire/dto/InquireReqDTO.java
@@ -1,0 +1,4 @@
+package com.example.umc10th.domain.inquire.dto;
+
+public class InquireReqDTO {
+}
