+++ Sangwan/src/main/java/com/example/umc10th/domain/inquire/enums/InquireStatus.java
@@ -1,4 +1,4 @@
-package com.example.umc10th.domain.qna.enums;
+package com.example.umc10th.domain.inquire.enums;
 
 public enum InquireStatus {
     WAITING, COMPLETED
