+++ Sangwan/src/main/java/com/example/umc10th/domain/inquire/enums/InquireType.java
@@ -1,0 +1,5 @@
+package com.example.umc10th.domain.inquire.enums;
+
+public enum InquireType {
+    SERVICE, STORE, PAYMENT, ACCOUNT, ETC
+}
