+++ Sangwan/src/main/java/com/example/umc10th/domain/inquire/repository/InquireRepository.java
@@ -1,0 +1,7 @@
+package com.example.umc10th.domain.inquire.repository;
+
+import com.example.umc10th.domain.inquire.entity.Inquire;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface InquireRepository extends JpaRepository<Inquire, Long> {
+}
