+++ Sangwan/src/main/java/com/example/umc10th/domain/inquire/service/InquireService.java
@@ -1,0 +1,4 @@
+package com.example.umc10th.domain.inquire.service;
+
+public class InquireService {
+}
