--- Sangwan/src/main/java/com/example/umc10th/domain/qna/controller/QnaController.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.qna.controller;
-
-public class QnaController {
-}
