--- Sangwan/src/main/java/com/example/umc10th/domain/qna/converter/QnaConverter.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.qna.converter;
-
-public class QnaConverter {
-}
