--- Sangwan/src/main/java/com/example/umc10th/domain/qna/dto/QnaResDTO.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.qna.dto;
-
-public class QnaResDTO {
-}
