--- Sangwan/src/main/java/com/example/umc10th/domain/qna/enums/InquireType.java
@@ -1,5 +1,0 @@
-package com.example.umc10th.domain.qna.enums;
-
-public enum InquireType {
-    // TODO: 문의 유형 추가
-}
