--- Sangwan/src/main/java/com/example/umc10th/domain/qna/repository/QnaRepository.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.qna.repository;
-
-public interface QnaRepository {
-}
