--- Sangwan/src/main/java/com/example/umc10th/domain/qna/service/QnaService.java
@@ -1,4 +1,0 @@
-package com.example.umc10th.domain.qna.service;
-
-public class QnaService {
-}
